Novas atividades

+++ LOGICA/Lista 9/Lista09_07.c
@@ -3,13 +3,13 @@
 int main (){
     int linha, i, j, k = 0, l = 10, g = 1;
     char op;
-    float m[12][12], soma = 0, media = 0;
+    double m[12][12], soma = 0, media = 0;
 
     scanf("%i", &linha);
     scanf(" %c", &op);
     for(i = 0; i < 12; i++){
         for(j = 0; j < 12; j++){
-            scanf("%f", &m[i][j]);
+            scanf("%lf", &m[i][j]);
         }
     }
     i = 0;
@@ -24,7 +24,7 @@ int main (){
             l--;
             g++;
         }
-        printf("%.1f\n", soma);
+        printf("%.1lf\n", soma);
     } else if(op == 'M'){
         for(i = 0; i < 12; i++){
             for(j = 0; j < 12; j++){
@@ -37,7 +37,7 @@ int main (){
             g++;       
         }
         media = soma / k;
-        printf("%.1f\n", media);
+        printf("%.1lf\n", media);
     }
     return 0;
 }
